requirements.txt
openai
yfinance
phidata
python-dotenv
lancedb
tantivy
pypdf
sqlalchemy
httpx
duckdb
duckduckgo-search
fastapi
uvicorn